planner_bot/pip-requirements.txt
django==3.2.3
python-telegram-bot==13.5